unbundled PySO8601 for security reasons
I have decided against bundling PySO8601 due to this conversation
https://github.com/j4mie/micromodels/pull/9#issuecomment-7457072

The primary reason I have decided to unbundle PySO8601 is to for
security reasons.  If a security flaw is found in PySO8601 and I have
not upgraded the bundled PySO8601, I subject developers to that bug
without their knowledge.  They may upgrade the PySO8601 package but
micromodels would still be vulnerable.

This is technically a backwards compatible change, so I will bump
the version to 0.6.0.